requirements.txt
pandas==1.5.3
scikit-learn==1.3.0
seaborn==0.12.2
matplotlib==3.7.1
mlflow==2.5.0
dvc==3.15.0
imbalanced-learn==0.11.0
scipy==1.11.1
kaggle==1.5.16
pytest==7.4.0
pyyaml==6.0
joblib==1.3.0
numpy==1.24.3
python-dotenv==1.0.0